# backend/api/openapi.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "invoice/v1/invoice.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/orders": {
      "get": {
        "operationId": "OrderService_ListOrders",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListOrdersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "OrderService"
        ]
      },
      "post": {
        "operationId": "OrderService_CreateOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateOrderRequest"
            }
          }
        ],
        "tags": [
          "OrderService"
        ]
      }
    },
    "/v1/orders/{id}": {
      "get": {
        "operationId": "OrderService_GetOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "OrderService"
        ]
      },
      "delete": {
        "operationId": "OrderService_DeleteOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "OrderService"
        ]
      }
    },
    "/v1/orders/{order.id}": {
      "put": {
        "operationId": "OrderService_UpdateOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "order.id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderServiceUpdateOrderBody"
            }
          }
        ],
        "tags": [
          "OrderService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "OrderServiceUpdateOrderBody": {
      "type": "object",
      "properties": {
        "order": {
          "type": "object",
          "properties": {
            "recipientFullName": {
              "type": "string"
            },
            "emailAddress": {
              "type": "string"
            },
            "deliveryAddress": {
              "type": "string"
            },
            "phone": {
              "type": "string"
            },
            "status": {
              "$ref": "#/definitions/v1OrderStatus"
            },
            "orderLines": {
              "type": "array",
              "items": {
                "type": "object",
                "$ref": "#/definitions/v1OrderLine"
              }
            }
          }
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1CreateOrderRequest": {
      "type": "object",
      "properties": {
        "order": {
          "$ref": "#/definitions/v1Order"
        }
      }
    },
    "v1CreateOrderResponse": {
      "type": "object",
      "properties": {
        "order": {
          "$ref": "#/definitions/v1Order"
        }
      }
    },
...
    "v1DeleteOrderResponse": {
      "type": "object"
    },
...
    "v1GetOrderResponse": {
      "type": "object",
      "properties": {
        "order": {
          "$ref": "#/definitions/v1Order"
        }
      }
    },
...
    "v1ListOrdersResponse": {
      "type": "object",
      "properties": {
        "orders": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Order"
          }
        }
      }
    },
...
    "v1Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "recipientFullName": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "deliveryAddress": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/v1OrderStatus"
        },
        "orderLines": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1OrderLine"
          }
        }
      }
    },
    "v1OrderLine": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "readOnly": true
        },
        "orderId": {
          "type": "string"
        },
        "productId": {
          "type": "string"
        },
        "price": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1OrderStatus": {
      "type": "string",
      "enum": [
        "ORDER_STATUS_UNSPECIFIED",
        "ORDER_STATUS_RECEIVED",
        "ORDER_STATUS_PROCESSING",
        "ORDER_STATUS_PICKING",
        "ORDER_STATUS_PACKED",
        "ORDER_STATUS_READY_SHIPMENT",
        "ORDER_STATUS_IN_TRANSIT",
        "ORDER_STATUS_DELIVERED",
        "ORDER_STATUS_CANCELLED",
        "ORDER_STATUS_ON_HOLD",
        "ORDER_STATUS_BACKORDERED",
        "ORDER_STATUS_RETURNED",
        "ORDER_STATUS_COMPLETED"
      ],
      "default": "ORDER_STATUS_UNSPECIFIED"
    },
...
    "v1UpdateOrderResponse": {
      "type": "object",
      "properties": {
        "order": {
          "$ref": "#/definitions/v1Order"
        }
      }
    },
...
  }
}